ai_agent/requirements.txt
fastapi
uvicorn
langchain
openai
pinecone-client
python-dotenv
langchain-community
langchain_openai